Store constants in program memory (ROM).
The macro definition "progmem" is used to declare constant values that
can be stored in program memory (i.e. ROM on the AVR).
This macro is an alias to the declaration of a section called ".progmem"
which is loaded into ROM (see linker.X scripts).
To retrieve the values stored in program memory during execution, the user
will call load_u8_from_progmem(char const * const source) (in startup.S) by
giving a pointer to the stored constant.
For the moment, this only works (only needed) with retrieving 8-bit values.