Added a way to provide task specific configuration when registering a task.
This configuration is contained in the struct TaskConfiguration which is passed
to Lz_Scheduler_RegisterTask(). This allows adding more configuration parameters
in the future without changing the proptotype of Lz_Scheduler_RegisterTask(), and
not overloading the stack with more function parameters.

If this new parameter is NULL, a default configuration will be applied.

The function Lz_Scheduler_InitTaskConfiguration() is provided to initialize the
struct TaskConfiguration with default values, so that the user can
change only the appropriate values without caring the others.