Setting up a static library containing the Lazuli kernel.
Now, the kernel code is first compiled but not linked. Then the .o files are
archived in a static library .a file (with ar command). This library contains
all .o files from kernel code and is then linked with user code to produce the
final ELF image containing all final executable code.

In the future the goal of all that is:
* To allow hiding the kernel API to the user code
* Maybe to allow distributing the kernel as a standalone binary static library