Manage task termination.
When a task returns from its main function, it is now excluded from
scheduling. We no longer need to put a for(;;); in the end of the main
function of a task. The result is a much more fluid execution. We can
clearly see that when executing the mutex example programs.

* Introduce a public API call Lz_Task_Terminate() to allow the user to
  terminate a task manually.
* Introduce a configuration option to set the saving of the context of
  the terminating task.
* Rename Lz_GetTaskName() to Lz_Task_GetName().
* Rename KMain() to Kernel_Main().
* Update example programs accordingly to take this automatic task
  termination management into consideration (i.e. remove all the
  unnecessary infinite loops).