Add the scheduling policy PRIORITY RR to the scheduler.
Until now, only cyclic tasks where managed by the scheduler, in an RMS fashion.
Now, the scheduler can handle a new scheduling policy for tasks: PRIORITY_RR.
This policy is the equivalent of POSIX's SCHED_RR.
The management of interrupts has been reviewed in order to respect time
constraints for cyclic realtime tasks.
Also define the macro DEPENDENCY_ON_MODULE(X) that can be used to stop
compilation if a needed module is not selected for linking.