CMake toolchain: Fix a bug involving the linking of static library and the propagation of compiler flags.
CMake toolchain upgraded to CMake version 3.12.

Object files of the kernel were re-included to the command line of final linking
although the static archive was already present. That was redundant. Now only the
static archive is linked to the user program.
Compiler flags for the kernel wrer propagated to the compilation of user program.
This also has been fixed.

Finally, the version string included on the final binary is no longer included
through a kernel module, but simply with a few lines in startup.S assembler file.