Variadic functions: Working implementation <stdarg.h>: of access to variadic function parameters.
This implementation relies on GCC builtins. This is contradictory with the
goal of implementing all of this project in pure C. However, the other
implementation shown in the file <stdarg.h>, which is pure C, suffers from severe
limitations. So for now we prefer to keep using this implementation as it
imposes less constraints to the final user.