Stack integrity: Introduce a tool to check the stack integrity on hand written ASM code.
This tool is quite rudimentary now, but does the job. For now it is executed
manually.

It checks if all the registers pushed onto the stack at the beginning of the
routine are pop'd in the correct order before returning.

It has some limitations. e.g. It cannot detect multiple returns in a single routine,
but the tool will improve when new needs arise.