printf: Change the inner "size" variable to be of type 'int'.
This is to conform to the return type of printf(), and it also allows to
bring the pad length to sizeof(int) too. Oddly, the resulting binary is slightly
smaller. And, more important, it doesn't increase the stack usage.
This also fixes a potential truncation bug when assigning the return value of strlen(),
which is 'int' to the previous "size", which was 'uint8_t'.